Improve caching of getHolidays
